* Made horizontal spacing parameter an "int" instead of a "float" since 	  "float" was making the fonts look blurry. Spacing changes will be 	  choppier, so we will need to revisit that.

+++ Headers/iTerm/PseudoTerminal.h
@@ -43,7 +43,7 @@
     
     /////////////////////////////////////////////////////////////////////////
     int WIDTH,HEIGHT;
-	float charWidth; // float for smoother increases
+	int charWidth;
 	int charHeight;
 	float charHorizontalSpacingMultiplier, charVerticalSpacingMultiplier;
     NSFont *FONT, *NAFONT;
